feat: add voice commands settings component and integrate into settings screen
