+++ src/components/settings/index.ts
@@ -1,5 +1,6 @@
 // src/components/settings/index.ts
 export { default as SettingsCheckBoxes } from './SettingsCheckBoxes';
+export { default as SettingsVoiceCommands } from './SettingsVoiceCommands';
 export { default as SettingsBackup } from './SettingsBackup';
 export { default as SettingsDataManagement } from './SettingsDataManagement';
 export { default as SettingsAppInfo } from './SettingsAppInfo';
